# docs/openapi.yaml
openapi: 3.1.0
info:
  title: dcbuilder.dev API
  description: REST API for managing jobs, candidates, investments, blog, and news content
  version: 1.0.0
  contact:
    name: dcbuilder
    url: https://dcbuilder.dev
servers:
  - url: /api/v1
    description: API v1
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
  schemas:
...
    BlogPost:
      type: object
      required:
        - slug
        - title
        - content
      properties:
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        title:
          type: string
        description:
          type: string
        content:
          type: string
          description: MDX content
        date:
          type: string
          format: date
        source:
          type: string
        sourceUrl:
          type: string
          format: uri
        image:
          type: string
          format: uri
        published:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
paths:
...
  /blog:
    get:
      tags: [Blog]
      summary: List blog posts
      description: Returns published posts for public access, all posts for authenticated users
      security:
        - {}
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of blog posts
    post:
      tags: [Blog]
      summary: Create blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPost'
      responses:
        '201':
          description: Blog post created
...